test: share registered background command dispatcher
